+++ src/utils.rs
@@ -1,0 +1,5 @@
+/// Helper function to calculate mid point
+#[doc(hidden)]
+pub fn midpoint(lhs: f64, rhs: f64) -> f64 {
+    (lhs + rhs) / 2.0
+}