+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.2.0 (May 15, 2025)
+
+- Added volume to the CandleStick trait
+- CandleStick trait object safe
+- Added typical price and money flow
+
 ## v0.1.1 (May 08, 2025)
 
 - Fixed docs.rs url for documentation